requirements.txt
Django==1.8.5
PyJWT==1.4.0
amqp==1.4.7
anyjson==0.3.3
billiard==3.3.0.21
celery==3.1.19
django-cors-headers==1.1.0
django-filter==0.11.0
djangorestframework==3.2.5
djangorestframework-filters==0.5.0
djangorestframework-gis==0.9.6
djangorestframework-jwt==1.7.2
httplib2==0.9.2
kombu==3.0.29
oauthlib==1.0.3
psycopg2==2.6.1
pytz==2015.7
redis==2.10.5
requests==2.8.1
requests-oauthlib==0.5.0
twython==3.3.0
